Add net-misc/network-manager-1.42.6-r2 with policy fix
The upstream ebuild by default sets the meson option "modify_system" to
"true" which causes the polkit policy installed for NetworkManager to
allow all users to modify system connections. That renders the part of
the ebuild that installs a polkit rule and dependency on the plugdev
group moot. That rule is supposed to allow only users in the plugdev
group to modify system connections. But if all users are by default
allowed to do it, then that rule and group have no effect.

This adds a USE flag "useradmin" (which I've defaulted on to match
current upstream behavior) and makes the meson "modify_system" option
depend on that USE flag. If disabled, then the policy defaults to not
allowing all users to modify system connections and the rule to allow
users in the plugdev group to modify them can be used to grant
permission to users that need it.