app-arch/xz-utils: Add old, possibly safe, version
Any version of XZ Utils after 5.2.12 has contributions that can't be
trusted. It could be the whole project can't be trusted, but right now
it's not feasible to just remove it. So going back to an older, possibly
safe, version seems the most prudent course of action at this point.