Add sys-kernel/genkernel with udev rule patch for new lvm2
Newer versions of genkernel use a newer version of lvm2 that comes with
a udev rules that causes errors during boot on systems that don't have
systemd. This provides a patch to genkernel, for patching lvm2 when it
builds it, so that the offending rule is modified to accomodate systems
that don't have systemd.