Add sys-kernel/linux-firmware-20240909-r1
Staring with linux-firmware-20241017-r3, symlinks in the nvidia tree are
broken causing firmware files to be left out (i.e. files listed in the
saved config that are symlinks aren't getting installed to
/lib/firmware).

sys-kernel/linux-firmware-20240909-r1 has been dropped upstream, so
adding it here so that I can stay on this old version, since there's so
far no sign of this getting fixed.